% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measures.R
\name{get_measures}
\alias{get_measures}
\title{Get Measures}
\usage{
get_measures(release)
}
\arguments{
\item{release}{The year of release for the requested data set (2020-2023)}
}
\value{
A viewer that opens a tibble of the available measures for a specified release year.
}
\description{
Use this function to source a list of available measures in the CDC PLACES data set.
}
\examples{

\dontrun{
get_measures(2022)
}

}
