% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measures22}
\alias{measures22}
\title{Measures included in the 2022 release of the CDC PLACES Data}
\usage{
data(measures22)
}
\description{
The primary data was obtained via the function get_places. Columns include
the year, the measureid, short_question_text (short name), measure (full desciption),
categoryid (the category of the measure), and release (the data release year).
}
