% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dictionary.R
\name{get_dictionary}
\alias{get_dictionary}
\title{View the 'CDC PLACES' data dictionary}
\usage{
get_dictionary()
}
\value{
a dataframe with the current PLACES data dictionary.
}
\description{
This function provides the user with a data frame that shows all of the available measures in the PLACES data set and for which release years the measures are included.
}
\examples{
# First save the data
dictionary <- get_dictionary()
# Then view the data frame
# View(dictionary)

}
